# [android] improve performance of `Entry.MaxLength`